wrapper/requirements.txt
httpx[http2]>=0.27.0
py-clob-client>=0.34.0
python-dotenv>=1.0.0
